// types/contact.ts
export type ContactContent = {
  title: string;
  subtitle: string;
  namePlaceholder: string;
  emailPlaceholder: string;
  messagePlaceholder: string;
  buttonLabel: string;
  email: string;
  facebook?: string;
  instagram?: string;
  snapchat?: string;
  github?: string;
  x?: string;
  linkedin?: string;
  website?: string;
};